third_party/blink/web_tests/external/wpt/mathml/presentation-markup/direction/direction-expected.txt

This is a testharness.js-based test.
FAIL Check direction assert_equals: expected "ltr" but got "rtl"
Harness: the test ran to completion.
